+++ Changelog.md
@@ -4,3 +4,8 @@
 ||||- Added results filtering.|
 ||||- Added highlighting of options selected in the map.|
 |0.1.1|Aug 19th, 2026|Mario Mora|- Fixed the responsive design for Mobile devices.|
+|0.2.0|Aug 20th, 2026|Mario Mora|- Add new component with roasters' available products.|
+||||- Add return to search button.|
+||||- Add roaster's name as title and address as subtitle.|
+||||- Add list of available products.|
+||||- Add an "Add to cart" button.|